infrastructure/serverless/azure-functions/requirements.txt
# Azure Functions
azure-functions>=1.18.0

# Azure SDK
azure-storage-blob>=12.19.0
azure-servicebus>=7.11.0
azure-eventgrid>=4.15.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.2
